+++ ice-20201109/src/Models/GetProjectExportJobRequest.php
@@ -1,0 +1,51 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\ICE\V20201109\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class GetProjectExportJobRequest extends Model
+{
+    /**
+     * @description This parameter is required.
+     *
+     * @example ****cdb3e74639973036bc84****
+     *
+     * @var string
+     */
+    public $jobId;
+    protected $_name = [
+        'jobId' => 'JobId',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->jobId) {
+            $res['JobId'] = $this->jobId;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return GetProjectExportJobRequest
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['JobId'])) {
+            $model->jobId = $map['JobId'];
+        }
+
+        return $model;
+    }
+}