+++ ice-20201109/ChangeLog.md
@@ -1,3 +1,15 @@
+2025-01-16 Version: 3.7.0
+- Support API GetProjectExportJob.
+- Support API SendAIAgentDataChannelMessage.
+- Support API SubmitProjectExportJob.
+- Support API SubmitScreenMediaHighlightsJob.
+- Support API TakeoverAIAgentCall.
+- Update API GetSmartHandleJob: update response param.
+- Update API ListSearchLib: update response param.
+- Update API QuerySearchLib: update response param.
+- Update API SubmitMediaAiAnalysisJob: add param UserData.
+
+
 2024-12-30 Version: 3.7.0
 - Support API SendAIAgentDataChannelMessage.
 - Support API SubmitScreenMediaHighlightsJob.